*** empty log message ***
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk/jasperreports@458 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ src/net/sf/jasperreports/charts/JRBar3DPlot.java
@@ -36,13 +36,13 @@ import net.sf.jasperreports.engine.JRExpression;
  * @version $Id$ 
  */
 public interface JRBar3DPlot extends JRChartPlot {
-    
-    public JRExpression getCategoryAxisLabelExpression();
-    
-    public JRExpression getValueAxisLabelExpression();
-    
-    public int getXOffset();
-    
-    public int getYOffset();
+	
+	public JRExpression getCategoryAxisLabelExpression();
+	
+	public JRExpression getValueAxisLabelExpression();
+	
+	public double getXOffset();
+	
+	public double getYOffset();
 
 }
